Added rnnt init of simple encoder
