+++ src/myrtlespeech/configs/rnnt_en.config
@@ -1,0 +1,67 @@
+speech_to_text {
+  alphabet: "_abcdefghijklmnopqrstuvwxyz '";
+
+  pre_process_step {
+    stage: TRAIN_AND_EVAL;
+    mfcc {
+      numcep: 83;
+      winlen: 0.025;
+      winstep: 0.010;
+    }
+  }
+
+  pre_process_step {
+    stage: TRAIN_AND_EVAL;
+    standardize {
+      mean: true;
+      reduce_dim: 1;
+    }
+  }
+
+
+  rnn_t {
+    rnn_t_encoder {
+      rnn1 {
+        rnn_type: LSTM;
+        hidden_size: 640;
+        num_layers: 3;
+        bias: true;
+        bidirectional: false;
+      }
+}
+}
+}
+
+  train_config {
+    batch_size: 64;
+    epochs: 1;
+    sgd {
+      learning_rate: 0.001;
+      momentum {
+        value: 0.9;
+      }
+      l2_weight_decay {
+        value: 0.0005;
+      }
+    }
+    dataset {
+      librispeech {
+        root: "~/Data/";
+        subset: TRAIN_CLEAN_100;
+        max_secs {
+          value: 16.7;
+        }
+      }
+    }
+    shuffle_batches_before_every_epoch: true;
+  }
+
+  eval_config {
+    batch_size: 64;
+    dataset {
+      librispeech {
+        root: "~/Data/";
+        subset: DEV_CLEAN;
+      }
+    }
+  }
