+++ src/myrtlespeech/protos/rnn_t_encoder.proto
@@ -1,0 +1,17 @@
+syntax = "proto3";
+
+package myrtlespeech.protos;
+
+import "google/protobuf/wrappers.proto";
+import "myrtlespeech/protos/rnn.proto";
+
+// RNN-T Encoder configuration
+message RNNTEncoder {
+  RNN rnn1 = 1;
+
+  uint32 time_reduction_factor = 3;
+
+  RNN rnn2 = 4;
+
+
+}
